requirements.txt
aiogram==3.0.0
sqlalchemy==2.0.0
asyncpg==0.27.0
aiosqlite==0.19.0
pydantic-settings==2.0.0
python-dotenv==1.0.0
beautifulsoup4==4.12.0
aiohttp==3.8.0